fixed missing preprocessing primary id type

+++ backend/src/main/java/com/bakdata/conquery/models/types/specific/MoneyTypeVarInt.java
@@ -15,7 +15,7 @@ public class MoneyTypeVarInt extends CTypeVarInt<Long> {
 
 	@JsonCreator
 	public MoneyTypeVarInt(VarIntType numberType) {
-		super(MajorTypeId.INTEGER, numberType);
+		super(MajorTypeId.MONEY, numberType);
 	}
 	
 	@Override
